`metric.go`: Use an atomic boolean for `initialized`.
Addresses gotsan-found data race when reading metric registration data over
urpc vs during initialization.

In practice, this can never happen, because the urpc call for metric
registration only happens after the boot sync file descriptor has been
written to, but there's no way for gotsan to know that. So this change just
switches it to an atomic boolean for good measure.

Also verified in `metric.GetSnapshot`, because why not.

PiperOrigin-RevId: 500043216